requirements.txt
# IBM watsonx.ai and AI/ML dependencies
ibm-watsonx-ai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
geopy>=2.4.0

# Web framework and API
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
httpx>=0.25.0

# Frontend and visualization
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.0.0
folium>=0.15.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database and storage
sqlalchemy>=2.0.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3